+++ experimental/packages/opentelemetry-instrumentation-http/test/functionals/http-enable.test.ts
@@ -13,8 +13,8 @@ import {
   DiagConsoleLogger,
   INVALID_SPAN_CONTEXT,
 } from '@opentelemetry/api';
-import { NodeTracerProvider } from '@opentelemetry/sdk-trace-node';
 import {
+  BasicTracerProvider,
   InMemorySpanExporter,
   SimpleSpanProcessor,
 } from '@opentelemetry/sdk-trace-base';
@@ -90,7 +90,7 @@ const hostname = 'localhost';
 const pathname = '/test';
 const serverName = 'my.server.name';
 const memoryExporter = new InMemorySpanExporter();
-const provider = new NodeTracerProvider({
+const provider = new BasicTracerProvider({
   spanProcessors: [new SimpleSpanProcessor(memoryExporter)],
 });
 instrumentation.setTracerProvider(provider);