+++ experimental/packages/opentelemetry-instrumentation-http/test/functionals/http-metrics.test.ts
@@ -8,7 +8,7 @@ import {
   InMemoryMetricExporter,
   MeterProvider,
 } from '@opentelemetry/sdk-metrics';
-import { NodeTracerProvider } from '@opentelemetry/sdk-trace-node';
+import { BasicTracerProvider } from '@opentelemetry/sdk-trace-base';
 import {
   ATTR_ERROR_TYPE,
   ATTR_HTTP_REQUEST_METHOD,
@@ -50,7 +50,7 @@ const serverPort = 22346;
 const protocol = 'http';
 const hostname = 'localhost';
 const pathname = '/test';
-const tracerProvider = new NodeTracerProvider();
+const tracerProvider = new BasicTracerProvider();
 const metricsMemoryExporter = new InMemoryMetricExporter(
   AggregationTemporality.DELTA
 );