+++ experimental/packages/opentelemetry-sdk-node/test/sdk.test.ts
@@ -29,13 +29,13 @@ import { OTLPMetricExporter as OTLPGrpcMetricExporter } from '@opentelemetry/exp
 import { OTLPMetricExporter as OTLPProtoMetricExporter } from '@opentelemetry/exporter-metrics-otlp-proto';
 import { OTLPMetricExporter as OTLPHttpMetricExporter } from '@opentelemetry/exporter-metrics-otlp-http';
 import { PrometheusExporter as PrometheusMetricExporter } from '@opentelemetry/exporter-prometheus';
-import { NodeTracerProvider } from '@opentelemetry/sdk-trace-node';
 import {
   assertServiceInstanceIdIsUUID,
   assertServiceResource,
 } from './util/resource-assertions';
 import type { IdGenerator, SpanProcessor } from '@opentelemetry/sdk-trace-base';
 import {
+  BasicTracerProvider,
   ConsoleSpanExporter,
   SimpleSpanProcessor,
   BatchSpanProcessor,
@@ -68,7 +68,10 @@ import { OTLPTraceExporter as OTLPProtoTraceExporter } from '@opentelemetry/expo
 import { OTLPTraceExporter as OTLPGrpcTraceExporter } from '@opentelemetry/exporter-trace-otlp-grpc';
 import { ZipkinExporter } from '@opentelemetry/exporter-zipkin';
 
-import { NOOP_COUNTER_METRIC } from '../../../../api/src/metrics/NoopMeter';
+import {
+  NOOP_COUNTER_METRIC,
+  NOOP_METER,
+} from '../../../../api/src/metrics/NoopMeter';
 import { ATTR_HOST_NAME, ATTR_PROCESS_PID } from '../src/semconv';
 
 function assertDefaultContextManagerRegistered() {
@@ -193,7 +196,7 @@ describe('Node SDK', () => {
       assert.strictEqual(setGlobalTracerProviderSpy.callCount, 1);
       assert.ok(
         setGlobalTracerProviderSpy.lastCall.args[0] instanceof
-          NodeTracerProvider
+          BasicTracerProvider
       );
       await sdk.shutdown();
     });
@@ -216,7 +219,7 @@ describe('Node SDK', () => {
       );
       assert.ok(
         setGlobalTracerProviderSpy.lastCall.args[0] instanceof
-          NodeTracerProvider
+          BasicTracerProvider
       );
       await sdk.shutdown();
     });
@@ -238,11 +241,13 @@ describe('Node SDK', () => {
       assertDefaultContextManagerRegistered();
       assertDefaultPropagatorRegistered();
 
-      const nodeTracerProvider = setGlobalTracerProviderSpy.lastCall.args[0];
+      const tracerProvider = setGlobalTracerProviderSpy.lastCall.args[0];
       assert.strictEqual(setGlobalTracerProviderSpy.callCount, 1);
-      assert.ok(nodeTracerProvider instanceof NodeTracerProvider);
+      assert.ok(tracerProvider instanceof BasicTracerProvider);
 
-      const spanProcessor = nodeTracerProvider['_activeSpanProcessor'] as any;
+      const spanProcessor = tracerProvider['_delegate'][
+        '_activeSpanProcessor'
+      ] as any;
 
       assert.ok(
         spanProcessor.constructor.name === 'MultiSpanProcessor',
@@ -434,9 +439,10 @@ describe('Node SDK', () => {
 
       assert.strictEqual(setGlobalTracerProviderSpy.callCount, 1);
       const tracerProvider = setGlobalTracerProviderSpy.lastCall.args[0];
-      assert.ok(tracerProvider instanceof NodeTracerProvider);
+      assert.ok(tracerProvider instanceof BasicTracerProvider);
       assert.ok(
-        (tracerProvider as any)._config.meterProvider instanceof MeterProvider
+        (tracerProvider as any)._delegate._config.meterProvider instanceof
+          MeterProvider
       );
 
       const loggerProvider = setGlobalLoggerProviderSpy.lastCall.args[0];
@@ -474,8 +480,11 @@ describe('Node SDK', () => {
 
       assert.strictEqual(setGlobalTracerProviderSpy.callCount, 1);
       const tracerProvider = setGlobalTracerProviderSpy.lastCall.args[0];
-      assert.ok(tracerProvider instanceof NodeTracerProvider);
-      assert.equal((tracerProvider as any)._config.meterProvider, undefined);
+      assert.ok(tracerProvider instanceof BasicTracerProvider);
+      assert.deepEqual(
+        (tracerProvider as any)._delegate._config.meterProvider.getMeter(),
+        NOOP_METER
+      );
 
       const loggerProvider = setGlobalLoggerProviderSpy.lastCall.args[0];
       assert.deepEqual(
@@ -1722,9 +1731,10 @@ describe('Node SDK', () => {
     const getSdkSpanProcessors = (sdk: NodeSDK) => {
       const tracerProvider = sdk['_tracerProvider'];
 
-      assert.ok(tracerProvider instanceof NodeTracerProvider);
+      assert.ok(tracerProvider instanceof BasicTracerProvider);
 
-      const activeSpanProcessor = tracerProvider['_activeSpanProcessor'];
+      const activeSpanProcessor =
+        tracerProvider['_delegate']['_activeSpanProcessor'];
 
       assert.ok(activeSpanProcessor.constructor.name === 'MultiSpanProcessor');
 
@@ -1815,7 +1825,8 @@ describe('Node SDK', () => {
       const listOfProcessors = getSdkSpanProcessors(sdk);
 
       assert.ok(
-        sdk['_tracerProvider']!['_config']?.sampler instanceof AlwaysOffSampler
+        sdk['_tracerProvider']!['_delegate']['_config']?.sampler instanceof
+          AlwaysOffSampler
       );
       assert.strictEqual(listOfProcessors.length, 1);
       assert.ok(listOfProcessors[0] instanceof SimpleSpanProcessor);