+++ packages/opentelemetry-sdk-trace-base/src/types.ts
@@ -56,6 +56,11 @@ export interface TracerConfig {
 }
 
 /**
+ * TracerProviderOptions provides an interface for configuring a Tracer Provider.
+ */
+export type TracerProviderOptions = TracerConfig;
+
+/**
  * Configuration options for registering the API with the SDK.
  * Undefined values may be substituted for defaults, and null
  * values will not be registered.