+++ packages/opentelemetry-sdk-trace-node/src/config.ts
@@ -7,5 +7,8 @@ import type { TracerConfig } from '@opentelemetry/sdk-trace-base';
 
 /**
  * NodeTracerConfig provides an interface for configuring a Node Tracer.
+ *
+ * @deprecated this interface and package will be removed in next major (3.0) and replaced by
+ * the `TracerProviderConfig` interface of the future `@opentelemetry/sdk-trace` package.
  */
 export type NodeTracerConfig = TracerConfig;